Move contrastiveVI model and module files to external
