+++ scvi/external/contrastivevi/__init__.py
@@ -1,0 +1,4 @@
+from ._model import ContrastiveVI
+from ._module import ContrastiveVAE
+
+__all__ = ["ContrastiveVI", "ContrastiveVAE"]
