+++ scvi/model/__init__.py
@@ -2,7 +2,6 @@ from . import utils
 from ._amortizedlda import AmortizedLDA
 from ._autozi import AUTOZI
 from ._condscvi import CondSCVI
-from ._contrastivevi import ContrastiveVI
 from ._destvi import DestVI
 from ._jaxscvi import JaxSCVI
 from ._linear_scvi import LinearSCVI
@@ -25,5 +24,4 @@ __all__ = [
     "AmortizedLDA",
     "utils",
     "JaxSCVI",
-    "ContrastiveVI",
 ]
