+++ scvi/module/__init__.py
@@ -1,7 +1,6 @@
 from ._amortizedlda import AmortizedLDAPyroModule
 from ._autozivae import AutoZIVAE
 from ._classifier import Classifier
-from ._contrastivevae import ContrastiveVAE
 from ._jaxvae import JaxVAE
 from ._mrdeconv import MRDeconv
 from ._multivae import MULTIVAE
@@ -24,5 +23,4 @@ __all__ = [
     "MULTIVAE",
     "AmortizedLDAPyroModule",
     "JaxVAE",
-    "ContrastiveVAE",
 ]
