+++ tests/module/test_contrastivevae.py
@@ -2,8 +2,8 @@ import pytest
 import torch
 
 from scvi.dataloaders import ContrastiveDataLoader
+from scvi.external.contrastivevi import ContrastiveVAE
 from scvi.model._utils import _init_library_size
-from scvi.module import ContrastiveVAE
 from scvi.module.base import LossOutput
 
 REQUIRED_DATA_SOURCES = ["background", "target"]
